Use types where possible

+++ src/services/user.service.ts
@@ -1,22 +1,23 @@
 import { User } from '@prisma/client';
 import NotFoundError from '../errors/NotFoundError';
 import prisma from '../prisma';
+import UserDTO from '../types/UserDTO';
 
-export const deleteUser = async (id: string): Promise<void> => {
+export const deleteUser = async (userId: string): Promise<void> => {
   try {
-    await prisma.user.delete({ where: { id } });
+    await prisma.user.delete({ where: { id: userId } });
   } catch (err) {
     throw new NotFoundError('User not found');
   }
 };
 
-export const createUser = async (data: User): Promise<User> => {
+export const createUser = async (data: UserDTO): Promise<User> => {
   const user = await prisma.user.create({ data });
   return user;
 };
 
-export const getUser = async (id: string): Promise<User | null> => {
-  const user = await prisma.user.findUnique({ where: { id } });
+export const getUser = async (userId: string): Promise<User> => {
+  const user = await prisma.user.findUnique({ where: { id: userId } });
   if (!user) {
     throw new NotFoundError('User not found');
   }
